# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LinkItem {
  id: string;
  label: string;
  url: string;
  emoji: string;
  order: number;
}

export interface Profile {
  photo: string;
  name: string;
  description: string;
}

export interface Design {
  bgColor: string;
  glowColor: string;
  glowOpacity: number;
  accentColor: string;
  textColor: string;
}

export type LinktreeDataClicks = { [key: string]: number };

export interface LinktreeData {
  profile: Profile;
  links: LinkItem[];
  design: Design;
  clicks: LinktreeDataClicks;
}

export interface SaveLinktreeRequest {
  profile: Profile;
  links: LinkItem[];
  design: Design;
}

export interface SaveResult {
  success: boolean;
}

export interface TrackResult {
  clicks: number;
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export interface ErrorEnvelope {
  error: string;
}