# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /linktree:
    get:
      operationId: getLinktreeData
      tags: [linktree]
      summary: Get profile, links, design and click data
      responses:
        "200":
          description: Linktree data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LinktreeData"
    post:
      operationId: saveLinktreeData
      tags: [linktree]
      summary: Save profile, links, and design
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveLinktreeRequest"
      responses:
        "200":
          description: Saved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaveResult"
  /linktree/track/{id}:
    post:
      operationId: trackClick
      tags: [linktree]
      summary: Track a link click
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated click count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackResult"
...
components:
  schemas:
...
    LinkItem:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        url:
          type: string
        emoji:
          type: string
        order:
          type: integer
      required:
        - id
        - label
        - url
        - emoji
        - order
    Profile:
      type: object
      properties:
        photo:
          type: string
        name:
          type: string
        description:
          type: string
      required:
        - photo
        - name
        - description
    Design:
      type: object
      properties:
        bgColor:
          type: string
        glowColor:
          type: string
        glowOpacity:
          type: number
        accentColor:
          type: string
        textColor:
          type: string
      required:
        - bgColor
        - glowColor
        - glowOpacity
        - accentColor
        - textColor
    LinktreeData:
      type: object
      properties:
        profile:
          $ref: "#/components/schemas/Profile"
        links:
          type: array
          items:
            $ref: "#/components/schemas/LinkItem"
        design:
          $ref: "#/components/schemas/Design"
        clicks:
          type: object
          additionalProperties:
            type: integer
      required:
        - profile
        - links
        - design
        - clicks
    SaveLinktreeRequest:
      type: object
      properties:
        profile:
          $ref: "#/components/schemas/Profile"
        links:
          type: array
          items:
            $ref: "#/components/schemas/LinkItem"
        design:
          $ref: "#/components/schemas/Design"
      required:
        - profile
        - links
        - design
    SaveResult:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
    TrackResult:
      type: object
      properties:
        clicks:
          type: integer
      required:
        - clicks
...