# lib/api-zod/src/generated/types/design.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Design {
  bgColor: string;
  glowColor: string;
  glowOpacity: number;
  accentColor: string;
  textColor: string;
}