# lib/api-zod/src/generated/types/linktreeData.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Design } from "./design";
import type { LinkItem } from "./linkItem";
import type { LinktreeDataClicks } from "./linktreeDataClicks";
import type { Profile } from "./profile";

export interface LinktreeData {
  profile: Profile;
  links: LinkItem[];
  design: Design;
  clicks: LinktreeDataClicks;
}