# lib/api-zod/src/generated/types/saveLinktreeRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Design } from "./design";
import type { LinkItem } from "./linkItem";
import type { Profile } from "./profile";

export interface SaveLinktreeRequest {
  profile: Profile;
  links: LinkItem[];
  design: Design;
}